Changed all Byte types to type Integer. Initially I thought this would save memory but after investigating it seems simpler to use Integer types as Java performs upcasting to integers for many operations and supposedly a lot of CPU architecture is built around fast integer operations.

+++ src/Dice.java
@@ -1,44 +1,59 @@
 public class Dice {
-    /** The number of sides of a die. */
-    public static final byte NUM_SIDES = 6;
+    /**
+     * The number of sides of a die.
+     */
+    public static final int NUM_SIDES = 6;
 
-    /** Constructs a pair of dice. Initially both dice are set to 0. */
+    /**
+     * Constructs a pair of dice. Initially both dice are set to 0.
+     */
     Dice() {
         _first = 0;
         _second = 0;
     }
 
-    /** Instantiate a pair of dice with values FIRST and SECOND, respectively. */
-    Dice(byte first, byte second) {
+    /**
+     * Instantiate a pair of dice with values FIRST and SECOND, respectively.
+     */
+    Dice(int first, int second) {
         _first = first;
         _second = second;
     }
 
 
-    /** Getter for the value of my first dice roll. */
-    public byte first() {
+    /**
+     * Getter for the value of my first dice roll.
+     */
+    public int first() {
         return _first;
     }
 
-    /** Getter for the value of my second dice roll. */
-    public byte second() {
+    /**
+     * Getter for the value of my second dice roll.
+     */
+    public int second() {
         return _second;
     }
 
-    /** Roll my dice and update the values. */
+    /**
+     * Roll my dice and update the values.
+     */
     public void roll() {
         _first = generateRoll();
         _second = generateRoll();
     }
 
 
-
-    /** Generate a random dice roll. Assumes a NUM_SIDES sided dice, where every integer between 1 and NUM_SIDES has equal (uniform) probability of occurring. */
-    static byte generateRoll() {
-        return (byte) Utils.randomIntegerInclusive(1, NUM_SIDES);
+    /**
+     * Generate a random dice roll. Assumes a NUM_SIDES sided dice, where every integer between 1 and NUM_SIDES has equal (uniform) probability of occurring.
+     */
+    static int generateRoll() {
+        return Utils.randomIntegerInclusive(1, NUM_SIDES);
     }
 
-    /** Return true iff the dice roll is a Pasch, that is the values of both rolls are equal. */
+    /**
+     * Return true iff the dice roll is a Pasch, that is the values of both rolls are equal.
+     */
     public boolean pasch() {
         return _first == _second;
     }
@@ -51,9 +66,13 @@ public class Dice {
         return "Roll: (" + _first + ", " + _second + ")";
     }
 
-    /** The value of my first roll. */
-    private byte _first;
+    /**
+     * The value of my first roll.
+     */
+    private int _first;
 
-    /** The value of my second roll. */
-    private byte _second;
+    /**
+     * The value of my second roll.
+     */
+    private int _second;
 }
